split elementwise_op.h into two header files
